+++ ormar/queryset/__init__.py
@@ -2,6 +2,7 @@
 Contains QuerySet and different Query classes to allow for constructing of sql queries.
 """
 from ormar.queryset.actions import FilterAction, OrderAction
+from ormar.queryset.clause import and_, or_
 from ormar.queryset.filter_query import FilterQuery
 from ormar.queryset.limit_query import LimitQuery
 from ormar.queryset.offset_query import OffsetQuery
@@ -16,4 +17,6 @@ __all__ = [
     "OrderQuery",
     "FilterAction",
     "OrderAction",
+    "and_",
+    "or_",
 ]