+++ ormar/queryset/actions/order_action.py
@@ -34,6 +34,31 @@ class OrderAction(QueryAction):
     def field_alias(self) -> str:
         return self.target_model.get_column_alias(self.field_name)
 
+    def get_field_name_text(self) -> str:
+        """
+        Escapes characters if it's required.
+        Substitutes values of the models if value is a ormar Model with its pk value.
+        Compiles the clause.
+
+        :return: complied and escaped clause
+        :rtype: sqlalchemy.sql.elements.TextClause
+        """
+        prefix = f"{self.table_prefix}_" if self.table_prefix else ""
+        return f"{prefix}{self.table}" f".{self.field_alias}"
+
+    def get_min_or_max(self) -> sqlalchemy.sql.expression.TextClause:
+        """
+        Used in limit sub queries where you need to use aggregated functions
+        in order to order by columns not included in group by.
+
+        :return: min or max function to order
+        :rtype: sqlalchemy.sql.elements.TextClause
+        """
+        prefix = f"{self.table_prefix}_" if self.table_prefix else ""
+        if self.direction == "":
+            return text(f"min({prefix}{self.table}" f".{self.field_alias})")
+        return text(f"max({prefix}{self.table}" f".{self.field_alias}) desc")
+
     def get_text_clause(self) -> sqlalchemy.sql.expression.TextClause:
         """
         Escapes characters if it's required.