+++ ormar/queryset/queryset.py
@@ -20,7 +20,7 @@ from ormar import MultipleMatches, NoMatch
 from ormar.exceptions import ModelError, ModelPersistenceError, QueryDefinitionError
 from ormar.queryset import FilterQuery
 from ormar.queryset.actions.order_action import OrderAction
-from ormar.queryset.clause import QueryClause
+from ormar.queryset.clause import FilterGroup, QueryClause
 from ormar.queryset.prefetch_query import PrefetchQuery
 from ormar.queryset.query import Query
 
@@ -192,6 +192,34 @@ class QuerySet:
             return self.model.merge_instances_list(result_rows)  # type: ignore
         return result_rows
 
+    def _resolve_filter_groups(self, groups: Any) -> List[FilterGroup]:
+        """
+        Resolves filter groups to populate FilterAction params in group tree.
+
+        :param groups: tuple of FilterGroups
+        :type groups: Any
+        :return: list of resolver groups
+        :rtype: List[FilterGroup]
+        """
+        filter_groups = []
+        if groups:
+            for group in groups:
+                if not isinstance(group, FilterGroup):
+                    raise QueryDefinitionError(
+                        "Only ormar.and_ and ormar.or_ "
+                        "can be passed as filter positional"
+                        " arguments,"
+                        "other values need to be passed by"
+                        "keyword arguments"
+                    )
+                group.resolve(
+                    model_cls=self.model,
+                    select_related=self._select_related,
+                    filter_clauses=self.filter_clauses,
+                )
+                filter_groups.append(group)
+        return filter_groups
+
     @staticmethod
     def check_single_result_rows_count(rows: Sequence[Optional["Model"]]) -> None:
         """
@@ -256,7 +284,9 @@ class QuerySet:
         # print("\n", exp.compile(compile_kwargs={"literal_binds": True}))
         return exp
 
-    def filter(self, _exclude: bool = False, **kwargs: Any) -> "QuerySet":  # noqa: A003
+    def filter(  # noqa: A003
+        self, *args: Any, _exclude: bool = False, **kwargs: Any
+    ) -> "QuerySet":
         """
         Allows you to filter by any `Model` attribute/field
         as well as to fetch instances, with a filter across an FK relationship.
@@ -268,6 +298,8 @@ class QuerySet:
         *  contains - like `album__name__contains='Mal'` (sql like)
         *  icontains - like `album__name__icontains='mal'` (sql like case insensitive)
         *  in - like `album__name__in=['Malibu', 'Barclay']` (sql in)
+        *  isnull - like `album__name__isnull=True` (sql is null)
+           (isnotnull `album__name__isnull=False` (sql is not null))
         *  gt - like `position__gt=3` (sql >)
         *  gte - like `position__gte=3` (sql >=)
         *  lt - like `position__lt=3` (sql <)
@@ -284,12 +316,14 @@ class QuerySet:
         :return: filtered QuerySet
         :rtype: QuerySet
         """
+        filter_groups = self._resolve_filter_groups(groups=args)
         qryclause = QueryClause(
             model_cls=self.model,
             select_related=self._select_related,
             filter_clauses=self.filter_clauses,
         )
         filter_clauses, select_related = qryclause.prepare_filter(**kwargs)
+        filter_clauses = filter_clauses + filter_groups  # type: ignore
         if _exclude:
             exclude_clauses = filter_clauses
             filter_clauses = self.filter_clauses
@@ -303,7 +337,7 @@ class QuerySet:
             select_related=select_related,
         )
 
-    def exclude(self, **kwargs: Any) -> "QuerySet":  # noqa: A003
+    def exclude(self, *args: Any, **kwargs: Any) -> "QuerySet":  # noqa: A003
         """
         Works exactly the same as filter and all modifiers (suffixes) are the same,
         but returns a *not* condition.
@@ -322,7 +356,7 @@ class QuerySet:
         :return: filtered QuerySet
         :rtype: QuerySet
         """
-        return self.filter(_exclude=True, **kwargs)
+        return self.filter(_exclude=True, *args, **kwargs)
 
     def select_related(self, related: Union[List, str]) -> "QuerySet":
         """